Make open_internal UTF-8-aware
open_internal was attempting to lookup a file's tuple by doing
resolve_cstring. However, resolve_cstring is not parsing UTF-8 strings
correctly; it works only on one byte at a time, and while push_character
is UTF-8-aware, it can't properly encode values given one byte at a
time.

It's not unthinkable that we might want to restrict some things to
ASCII-only. So this commit adds an UTF-8 counterpart called
resolve_utf8, and amends open_internal to use it.

Closes #567.

Signed-off-by: Alexandru Lazar <alazar@startmail.com>